+++ src/controller/model/response/userResponse.go
@@ -5,5 +5,5 @@ type UserResponse struct {
 	Username string `json:"username"`
 	Email    string `json:"email"`
 	Name     string `json:"name"`
-	Age      int    `json:"age"`
+	Age      int8   `json:"age"`
 }
